requirements.txt
asgiref==3.5.2
attrs==19.3.0
coverage==5.1
deepdiff==6.2.1
Django==4.1.2
django-fakery==2.3.0
djangorestframework==3.14.0
exceptiongroup==1.0.0
Faker==4.18.0
importlib-metadata==1.6.0
iniconfig==1.1.1
more-itertools==8.3.0
ordered-set==4.1.0
packaging==21.3
pluggy==0.13.1
psycopg2==2.9.5
py==1.9.0
pyparsing==3.0.9
pytest==6.2.5
pytest-cov==2.9.0
pytest-django==3.9.0
python-dateutil==2.8.2
python-dotenv==0.14.0
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
wcwidth==0.2.5
zipp==3.4.0
